catched all coin api bugs
