+++ src/core/coin/binance.py
@@ -17,6 +17,7 @@ from src.core.coin.lib.binance_api.exceptions import (BinanceAPIException,
                                                       BinanceRequestException,
                                                       BinanceWithdrawException)
 from src.core.util.exceptions import BinanceException
+from requests.exceptions import ReadTimeout, ConnectionError
 
 urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
 
@@ -47,7 +48,8 @@ class Binance(Coin):
         try:
             res = self._client.get_server_time()  # UTC Zone UnixStamp
             return res["serverTime"]
-        except (KeyError, BinanceAPIException, BinanceRequestException,
+        except (ReadTimeout, ConnectionError, KeyError,
+                BinanceAPIException, BinanceRequestException,
                 BinanceOrderException, BinanceWithdrawException) as err:
             raise BinanceException(err)
 
@@ -70,7 +72,8 @@ class Binance(Coin):
                 "webSockets_second": ''
             }
             return res
-        except (KeyError, BinanceAPIException, BinanceRequestException,
+        except (ReadTimeout, ConnectionError, KeyError,
+                BinanceAPIException, BinanceRequestException,
                 BinanceOrderException, BinanceWithdrawException) as err:
             raise BinanceException(err)
 
@@ -84,7 +87,8 @@ class Binance(Coin):
                 tSymbol = b["quoteAsset"]
                 res.append({"fSymbol": fSymbol, "tSymbol": tSymbol})
             return res
-        except (KeyError, BinanceAPIException, BinanceRequestException,
+        except (ReadTimeout, ConnectionError, KeyError,
+                BinanceAPIException, BinanceRequestException,
                 BinanceOrderException, BinanceWithdrawException) as err:
             raise BinanceException(err)
 
@@ -154,7 +158,8 @@ class Binance(Coin):
                     "min_notional": min_notional
                 })
             return res
-        except (KeyError, BinanceAPIException, BinanceRequestException,
+        except (ReadTimeout, ConnectionError, KeyError,
+                BinanceAPIException, BinanceRequestException,
                 BinanceOrderException, BinanceWithdrawException) as err:
             raise BinanceException(err)
 
@@ -174,7 +179,8 @@ class Binance(Coin):
                 "ask_one_size": ticker["askQty"]
             }
             return res
-        except (KeyError, BinanceAPIException, BinanceRequestException,
+        except (ReadTimeout, ConnectionError, KeyError,
+                BinanceAPIException, BinanceRequestException,
                 BinanceOrderException, BinanceWithdrawException) as err:
             raise BinanceException(err)
 
@@ -194,7 +200,8 @@ class Binance(Coin):
                 "ask_price_size": ticker["asks"]
             }
             return res
-        except (KeyError, BinanceAPIException, BinanceRequestException,
+        except (ReadTimeout, ConnectionError, KeyError,
+                BinanceAPIException, BinanceRequestException,
                 BinanceOrderException, BinanceWithdrawException) as err:
             raise BinanceException(err)
 
@@ -236,7 +243,8 @@ class Binance(Coin):
                     "volume": k[5]
                 })
             return res
-        except (KeyError, BinanceAPIException, BinanceRequestException,
+        except (ReadTimeout, ConnectionError, KeyError,
+                BinanceAPIException, BinanceRequestException,
                 BinanceOrderException, BinanceWithdrawException) as err:
             raise BinanceException(err)
 
@@ -245,7 +253,8 @@ class Binance(Coin):
         try:
             res = self._client.get_trade_fee(**kwargs)
             return res["tradeFee"]
-        except (KeyError, BinanceAPIException, BinanceRequestException,
+        except (ReadTimeout, ConnectionError, KeyError,
+                BinanceAPIException, BinanceRequestException,
                 BinanceOrderException, BinanceWithdrawException) as err:
             raise BinanceException(err)
 
@@ -290,7 +299,8 @@ class Binance(Coin):
                     float(ratio) * float(item["cummulativeQuoteQty"])
                 })
             return res
-        except (KeyError, BinanceAPIException, BinanceRequestException,
+        except (ReadTimeout, ConnectionError, KeyError,
+                BinanceAPIException, BinanceRequestException,
                 BinanceOrderException, BinanceWithdrawException) as err:
             raise BinanceException(err)
 
@@ -338,7 +348,8 @@ class Binance(Coin):
                     float(ratio) * float(item["cummulativeQuoteQty"])
                 })
             return res
-        except (KeyError, BinanceAPIException, BinanceRequestException,
+        except (ReadTimeout, ConnectionError, KeyError,
+                BinanceAPIException, BinanceRequestException,
                 BinanceOrderException, BinanceWithdrawException) as err:
             raise BinanceException(err)
 
@@ -385,7 +396,8 @@ class Binance(Coin):
                         float(ratio) * float(item["cummulativeQuoteQty"])
                     })
             return res
-        except (KeyError, BinanceAPIException, BinanceRequestException,
+        except (ReadTimeout, ConnectionError, KeyError,
+                BinanceAPIException, BinanceRequestException,
                 BinanceOrderException, BinanceWithdrawException) as err:
             raise BinanceException(err)
 
@@ -402,7 +414,8 @@ class Binance(Coin):
                     "locked": float(b["locked"])
                 })
             return res
-        except (KeyError, BinanceAPIException, BinanceRequestException,
+        except (ReadTimeout, ConnectionError, KeyError,
+                BinanceAPIException, BinanceRequestException,
                 BinanceOrderException, BinanceWithdrawException) as err:
             raise BinanceException(err)
 
@@ -419,7 +432,8 @@ class Binance(Coin):
                     "min_withdraw": float(value["minWithdrawAmount"])
                 })
             return res
-        except (KeyError, BinanceAPIException, BinanceRequestException,
+        except (ReadTimeout, ConnectionError, KeyError,
+                BinanceAPIException, BinanceRequestException,
                 BinanceOrderException, BinanceWithdrawException) as err:
             raise BinanceException(err)
 
@@ -434,7 +448,8 @@ class Binance(Coin):
                 "locked": float(base["locked"])
             }
             return res
-        except (KeyError, BinanceAPIException, BinanceRequestException,
+        except (ReadTimeout, ConnectionError, KeyError,
+                BinanceAPIException, BinanceRequestException,
                 BinanceOrderException, BinanceWithdrawException) as err:
             raise BinanceException(err)
 
@@ -448,7 +463,8 @@ class Binance(Coin):
                 "withdraw": withdraw["withdrawList"]
             }
             return res
-        except (KeyError, BinanceAPIException, BinanceRequestException,
+        except (ReadTimeout, ConnectionError, KeyError,
+                BinanceAPIException, BinanceRequestException,
                 BinanceOrderException, BinanceWithdrawException) as err:
             raise BinanceException(err)
 
@@ -498,7 +514,8 @@ class Binance(Coin):
                 "fee": float(ratio) * float(base["cummulativeQuoteQty"])
             }
             return res
-        except (KeyError, BinanceAPIException, BinanceRequestException,
+        except (ReadTimeout, ConnectionError, KeyError,
+                BinanceAPIException, BinanceRequestException,
                 BinanceOrderException, BinanceWithdrawException) as err:
             raise BinanceException(err)
 
@@ -531,7 +548,8 @@ class Binance(Coin):
                 "fee": float(ratio) * float(base["cummulativeQuoteQty"])
             }
             return res
-        except (KeyError, BinanceAPIException, BinanceRequestException,
+        except (ReadTimeout, ConnectionError, KeyError,
+                BinanceAPIException, BinanceRequestException,
                 BinanceOrderException, BinanceWithdrawException) as err:
             raise BinanceException(err)
 
@@ -547,7 +565,8 @@ class Binance(Coin):
             else:
                 res = {"order_id": orderID, "status": info["status"].lower()}
             return res
-        except (KeyError, BinanceAPIException, BinanceRequestException,
+        except (ReadTimeout, ConnectionError, KeyError,
+                BinanceAPIException, BinanceRequestException,
                 BinanceOrderException, BinanceWithdrawException) as err:
             raise BinanceException(err)
 
@@ -569,7 +588,8 @@ class Binance(Coin):
                         "status": info["status"].lower()
                     })
             return res
-        except (KeyError, BinanceAPIException, BinanceRequestException,
+        except (ReadTimeout, ConnectionError, KeyError,
+                BinanceAPIException, BinanceRequestException,
                 BinanceOrderException, BinanceWithdrawException) as err:
             raise BinanceException(err)
